% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose-seasonal.R
\name{stat_decomp}
\alias{stat_decomp}
\title{Classical seasonal adjustment Stat}
\usage{
stat_decomp(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  frequency = NULL,
  type = c("additive", "multiplicative"),
  index.ref = NULL,
  index.basis = 100,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. If specified
and \code{inherit.aes = TRUE} (the default), it is combined with the default
mapping at the top level of the plot. You must supply mapping if there is
no plot mapping.}

\item{data}{The data to be displayed in this layer. If NULL, this is
inherited in the call to \code{\link[ggplot2]{ggplot}}.}

\item{geom}{The geometric object to use to display the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never
includes, and \code{TRUE} always includes. It can also be a named logical vector
to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}

\item{frequency}{The frequency for the time series}

\item{type}{The type of seasonal component}

\item{index.ref}{if not NULL, a vector of integers indicating which elements
of the beginning of each series to use as a reference point for converting
to an index. If NULL, no conversion takes place and the data are presented
on the original scale.}

\item{index.basis}{if index.ref is not NULL, the basis point for converting
to an index, most commonly 100 or 1000.  See examples.}

\item{...}{other arguments for the geom}
}
\description{
Conducts seasonal adjustment on the fly for ggplot2, from classical seasonal
decomposition by moving averages
}
\details{
Classical decomposition is a very basic way of performing seasonal
  adjustment and is not recommended if you have access to X13-SEATS-ARIMA
  (\code{stat_seas}).  \code{stat_decomp} cannot allow the seasonality to
  vary over time, or take outliers into account in calculating seasonality.
}
\examples{
ap_df <- tsdf(AirPassengers)

# Default additive decomposition (doesn't work well in this case!):
ggplot(ap_df, aes(x = x, y = y)) +
   stat_decomp()

# Multiplicative decomposition, more appropriate:
ggplot(ap_df, aes(x = x, y = y)) +
   stat_decomp(type = "multiplicative")

# Multiple time series example:
ggplot(ldeaths_df, aes(x = YearMon, y = deaths, colour = sex)) +
  geom_point() +
  facet_wrap(~sex) +
  stat_decomp() +
  ggtitle("Seasonally adjusted lung deaths")

# Example using index:
ggplot(ldeaths_df, aes(x = YearMon, y = deaths, colour = sex)) +
  facet_wrap(~sex) +
  stat_decomp(index.ref = 1:12, index.basis = 1000) +
  ggtitle("Rolling annual median lung deaths, indexed (average month in 1974 = 1000)")

}
\seealso{
\code{\link[=stats]{decompose}}

Other time series stats for ggplot2: 
\code{\link{stat_index}()},
\code{\link{stat_rollapplyr}()},
\code{\link{stat_seas}()},
\code{\link{stat_stl}()}
}
\concept{time series stats for ggplot2}
