% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_rollmean.R
\name{stat_rollmean}
\alias{stat_rollmean}
\title{Rolling Mean Stat}
\usage{
stat_rollmean(mapping = NULL, data = NULL, geom = "line",
  position = "identity", show.legend = NA, inherit.aes = TRUE, width,
  align = "right", ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{width}{The width}

\item{align}{specifies whether the transformed series should be left or 
right-aligned or centered compared to the rolling window of observations}

\item{...}{other arguments for the geom}
}
\description{
Calculates rolling mean on the fly for ggplot2
}
\details{
blah blah blah
}
\examples{
ap_df <- tsdf(AirPassengers)

ggplot(ap_df, aes(x = x, y = y)) +
   stat_rollmean(width = 12)

ggplot(ldeaths_df, aes(x = YearMon, y = deaths, colour = sex)) +
  geom_point() +
  facet_wrap(~sex) +
  stat_rollmean(width = 12) +
  ggtitle("Seasonally adjusted lung deaths")

}
\seealso{
\code{\link{decompose}}

Other time.series.stats.for.ggplot2: \code{\link{stat_decomp}},
  \code{\link{stat_seas}}, \code{\link{stat_stl}}
}

