% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl-seasonal.R
\name{stat_stl}
\alias{stat_stl}
\title{LOESS seasonal adjustment Stat}
\usage{
stat_stl(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  frequency = NULL,
  s.window,
  index.ref = NULL,
  index.basis = 100,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. If specified
and \code{inherit.aes = TRUE} (the default), it is combined with the default
mapping at the top level of the plot. You must supply mapping if there is
no plot mapping.}

\item{data}{The data to be displayed in this layer. If NULL, this is
inherited in the call to \code{\link[ggplot2]{ggplot}}.}

\item{geom}{The geometric object to use to display the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never
includes, and \code{TRUE} always includes. It can also be a named logical vector
to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}

\item{frequency}{The frequency for the time series}

\item{s.window}{either the character string \code{"periodic"} or the span (in lags) of the 
loess window for seasonal extraction, which should be odd and at least 7, according to
Cleveland et al.  This has no default and must be chosen.}

\item{index.ref}{if not NULL, a vector of integers indicating which elements of
the beginning of each series to use as a reference point for converting to an index.  
If NULL, no conversion takes place and the data are presented on the original scale.}

\item{index.basis}{if index.ref is not NULL, the basis point for converting
to an index, most commonly 100 or 1000.  See examples.}

\item{...}{other arguments for the geom}
}
\description{
Conducts seasonal adjustment on the fly for ggplot2, from LOESS seasonal decomposition
}
\examples{
ap_df <- tsdf(AirPassengers)

# periodic if fixed seasonality; doesn't work well:
ggplot(ap_df, aes(x = x, y = y)) +
   stat_stl(s.window = "periodic")

# seasonality varies a bit over time, works better:
ggplot(ap_df, aes(x = x, y = y)) +
   stat_stl(s.window = 7)

# Multiple time series example:
ggplot(ldeaths_df, aes(x = YearMon, y = deaths, colour = sex)) +
   geom_point() +
   facet_wrap(~sex) +
   stat_stl(s.window = 7) +
   ggtitle("Seasonally adjusted lung deaths")

# Index so first value is 100:
ggplot(ap_df, aes(x = x, y = y)) +
   stat_stl(s.window = 7, index.ref = 1)
}
\seealso{
Other time series stats for ggplot2: 
\code{\link{stat_decomp}()},
\code{\link{stat_index}()},
\code{\link{stat_rollapplyr}()},
\code{\link{stat_seas}()}
}
\concept{time series stats for ggplot2}
