% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsdc.R
\name{ggsdc}
\alias{ggsdc}
\title{Visualise seasonal decomposition}
\usage{
ggsdc(data, mapping, frequency, method = c("stl", "decompose", "seas"), start,
  s.window, type = c("additive", "multiplicative"))
}
\arguments{
\item{data}{dataset to use for plot.}

\item{mapping}{List of aesthetic mappings.  Must include x and y, and optionally
can include colour/color}

\item{frequency}{frequency of the period of the time series eg 12 = monthly}

\item{method}{function to use for performing the seasonal decomposition. stl
and decompose are functions in the \code{stats} package; \code{seas} is access
to the \code{seats} program from X-13-SEATS-ARIMA via the \code{seasonal} package}

\item{start}{starting time for the data; only needed if \code{method = 'seas'}.}

\item{s.window}{parameter to pass to \code{stl()}}

\item{type}{parameter to pass to \code{decompose()}}
}
\value{
an object of class ggplot
}
\description{
Creates a four-facet plot of seasonal decomposition showing observed, trend, seasonal and random components
}
\details{
This function takes a data frame and performs seasonal decomposition
on the variable mapped to the y aesthetic, grouped by the variable (if any)
mapped to the colour or color aesthetic.  This allows the user to perform
the equivalent of plot(stats::decompose(x)) but in the ggplot2 environment for themes,
polishing etc; and to overlay decompositions on the same graphic; and with the 
X13-SEATS-ARIMA seasonal decomposition (so far only with default settings).
}
\examples{
# sample time series data in data frame
ap_df <- tsdf(AirPassengers)

ggsdc(ap_df, aes(x = x, y = y), method = "decompose", frequency = 12) +
   geom_line()
   
ggsdc(ap_df, aes(x = x, y = y), method = "decompose", frequency = 12, 
      type = "multiplicative") +
   geom_line(colour = "blue", size = 2) +
   theme_light(8)

ggsdc(ap_df, aes(x = x, y = y), method = "stl", frequency = 12, s.window = 7) +
   labs(x = "", y = "Air passenger numbers") +
   geom_point()
   
ggsdc(ldeaths_df, aes(x = YearMon, y = deaths, colour = sex), method = "seas", 
      frequency = 12, start = c(1949, 1)) +
      geom_line()
}
\seealso{
\code{\link{decompose}}, \code{\link{stl}}, \code{\link{seas}}
}

