% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{ldeaths_df}
\alias{ldeaths_df}
\title{Monthly Deaths from Lung Diseases in the UK}
\format{A data frame with 141 rows and 3 variables.}
\source{
P. J. Diggle (1990) \emph{Time Series: A Biostatistical Introduction}. Oxford, table A.3
}
\usage{
ldeaths_df
}
\description{
A long form combination of \code{fdeaths} and \code{mdeaths} from the \code{datasets} package.
}
\details{
\itemize{
  \item YearMon. Approximate, regular decimal representation of the beginning 
  of the period of measurement.  January 1974 is 1974.000
  \item sex.
  \item deaths. Monthly deaths from bronchitis, emphysema and asthma.
  ...
}
}
\seealso{
\code{\link{ldeaths}}
}
\keyword{datasets}

