% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{nzbop}
\alias{nzbop}
\title{New Zealand Balance of Payments major components 1971Q2 to 2015Q2}
\format{An object of class \code{data.frame} with 3676 rows and 5 columns.}
\source{
Statistics New Zealand \url{http://www.stats.govt.nz/browse_for_stats/economic_indicators/balance_of_payments/info-releases.aspx}
}
\usage{
nzbop
}
\description{
New Zealand's "BPM6 Quarterly, Balance of payments major components (Qrtly-Mar/Jun/Sep/Dec)".
}
\details{
"BPM6" refers to the sixth edition of the IMF's Balance of Payments and
International Investment Position Manual, which is the method used by Statistics
New Zealand to prepare these data.

Note:
\itemize{
  \item 'Value' is in millions of New Zealand dollars and is not adjusted for inflation.
  \item 'fob' means 'free on board'.
  \item 'inv.' stands for investment
  \item TimePeriod is the last day of the quarterly reference period ie \code{1971-06-30}
  means the fourth, fifth and six months of 1971.
}

This dataset was downloaded from \url{http://www.stats.govt.nz/infoshare/} and 
transformed in the following way:

\itemize{
  \item missing values were filtered out (ie of series that started later 
  than the longest series)
  \item a 'Balance' indicator variable was added for easier manipulation and filtering
  \item the single variable categorisation was split into two (Account
   and Category) to make it tidier.
  }
}
\keyword{datasets}

