% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{nzbop}
\alias{nzbop}
\title{New Zealand Balance of Payments selected variables 1972 to 2015}
\format{An object of class \code{data.frame} with 6063 rows and 5 columns.}
\source{
Statistics New Zealand \url{http://www.stats.govt.nz/browse_for_stats/economic_indicators/balance_of_payments/info-releases.aspx}
}
\usage{
nzbop
}
\description{
Selected exports and imports series from New Zealand's "BPM6 Quarterly (year ended in quarter), Balance of 
payments selected series (Qrtly-Mar/Jun/Sep/Dec)".
}
\details{
"BPM6" refers to the sixth edition of the IMF's Balance of Payments and
International Investment Position Manual, which is the method used by Statistics
New Zealand to prepare these data.

Note:
\itemize{
  \item 'Value' is in millions of New Zealand dollars and is not adjusted for inflation.
  \item 'nei' means 'not elsewhere included'.
  \item 'Category' items are not mutually exclusive; for example, 'Travel' is a 
  superset of 'Travel; Business'.
  \item TimePeriod is the last day of the reference period ie \code{1972-03-31}
  means the first three months of 1972.
}

This dataset was downloaded from \url{http://www.stats.govt.nz/infoshare/} and 
transformed in the following way:

\itemize{
  \item missing values were filtered out (ie of series that started later 
  than the longest series)
  \item estimates of balance were removed, leaving only exports and imports
  \item the single variable categorisation was split into three (Sector, Direction
   and Category) in accordance with 'tidy data' principles
  }
}
\keyword{datasets}

