% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13-seasonal.R
\name{stat_seas}
\alias{stat_seas}
\title{X13 seasonal adjustment Stat}
\usage{
stat_seas(mapping = NULL, data = NULL, geom = "line",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  x13_params = NULL, index.ref = NULL, index.basis = 100,
  frequency = NULL, start = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{x13_params}{a list of other parameters for \code{seas}}

\item{index.ref}{if not NULL, a vector of integers indicating which elements of
the beginning of each series to use as a reference point for converting to an index.  
If NULL, no conversion takes place and the data are presented on the original scale.}

\item{index.basis}{if index.ref is not NULL, the basis point for converting
to an index, most commonly 100 or 1000.  See examples.}

\item{frequency}{The frequency for the time series}

\item{start}{The starting point for the time series, in a format suitable for \code{ts()}}

\item{...}{other arguments for the geom}
}
\description{
Conducts X13-SEATS-ARIMA seasonal adjustment on the fly for ggplot2
}
\examples{
\dontrun{
ap_df <- tsdf(AirPassengers)

# SEATS with defaults:
ggplot(ap_df, aes(x = x, y = y)) +
   stat_seas()
   
# X11 with no outlier treatment:
ggplot(ap_df, aes(x = x, y = y)) +
  stat_seas(x13_params = list(x11 = "", outlier = NULL))

# Multiple time series example:    
ggplot(ldeaths_df, aes(x = YearMon, y = deaths, colour = sex)) +
  geom_point() +
  facet_wrap(~sex) +
  stat_seas() +
  ggtitle("Seasonally adjusted lung deaths")
  
# example use of index:  
ggplot(ap_df, aes(x = x, y = y)) +
  stat_seas(x13_params = list(x11 = "", outlier = NULL),
  index.ref = 1, index.basis = 1000) +
  labs(y = "Seasonally adjusted index\\n(first observation = 1000)")
  
# if the x value is not a decimal eg not created with time(your_ts_object),
# you need to specify start and frequency by hand:
ggplot(filter(nzbop, Account == "Current account"), 
      aes(x = TimePeriod, y = Value)) +
   stat_seas(start = c(1971, 2), frequency = 12) +
   facet_wrap(~Category, scales = "free_y")
  
  }
}
\seealso{
\code{\link{seas}}

Other time series stats for ggplot2: \code{\link{stat_decomp}},
  \code{\link{stat_index}}, \code{\link{stat_rollapplyr}},
  \code{\link{stat_stl}}
}
